+++ ios/RCTConvert+WebRTC.h
@@ -20,7 +20,8 @@ NS_ASSUME_NONNULL_BEGIN
 + (nullable RTCIceServer *)RTCIceServer:(nullable id)json;
 + (nullable RTCConfiguration *)RTCConfiguration:(nullable id)json;
 + (nullable WebRTCMediaStreamConstraints *)WebRTCMediaStreamConstraints:(nullable id)json;
-
++ (nullable RTCDataChannelConfiguration *)RTCDataChannelConfiguration:(nullable id)json;
++ (nullable RTCDataBuffer *)RTCDataBuffer:(nullable id)json;
 @end
 
 NS_ASSUME_NONNULL_END