+++ CHANGES.md
@@ -9,7 +9,16 @@
 - FIX
     - バグ修正
 
-## develop
+## 2020.3.0
+
+- [ADD] DataChannel に対応する
+    - @kdxu
+- [ADD] Android の onRemoveTrack に対応
+    - @enm10k
+- [CHANGE] CircleCI の設定を削除する
+    - @kdxu
+- [CHANGE]  サポートする React Native の最低バージョンを 0.61.5 に更新
+    - @enm10k
 
 ## 2020.2.0